move defines
